Getting started on recipes/quickstart.
NOTE: this version depends on prior
https://github.com/probcomp/bdbcontrib/pull/70

There are other independent changes here that I'm also happy to do in prior prs, but I wanted to start asking for feedback now, and continue expanding tests, and do prior prs as needed.

For example, one easily separable piece that I might want to make more general and put in bayeslite is the concept of a logger. Bayeslite uses the warnings module, which is great if that's what you want, but this format is more testable and more easily extensible for new interfaces like ipython.