Add keygen.py program and README.
As noted in the README, we need to discuss how we want to scale this
up if we do.  In particular, rather than add many public keys, we may
want to have a simple-minded certificate scheme instead.  Or we may
want to throw this whole thing out and forget an auto-downloadable
demo concept.